data-enrichment/1-artefact-color/pyproject.toml
[tool.poetry]
name = "color-extractor-service"
version = "0.2.2" # placeholder | version is defined in git tags | command: poetry version
description = "extracts fore- and background color palettes of images, including shiny colors (high satuated) "
authors = ["Nadav Babei <nbabai@3pc.de>, Sören Räuchle <sraeuchle@3pc.de>"]
license = "MIT"
readme = "README.md"

# https://pypi.org/project/poetry-dynamic-versioning/
[tool.poetry-dynamic-versioning]
enable = false
vcs = 'git'
style = "semver"
pattern = '(?x)^v*((?P<epoch>\d+)!)?(?P<base>\d+(\.\d+)*)' # 0.0.0 || v0.0.0 (git tag)

[tool.poetry.build]
generate-setup-file = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "1.26.1"
pillow = "10.1.0"
pandas = "2.1.2"
openpyxl = "3.1.2"
webcolors = "1.13"

torch =  { version = "2.1.0", platform='linux', source="pytorch-cu118" }
torchvision = { version = "0.16.0", platform='linux', source="pytorch-cu118" }
scikit-image = "0.22.0"
3pc-py-config = { version = "1.1.1", source="pypi-dreipc" }
pydantic = "2.4.2"
tqdm = "4.66.1"
scikit-learn="1.3.2"
fast_colorthief="0.0.4"

[[tool.poetry.source]]
name = 'pypi-dreipc'
url = 'https://nexus.3pc.de/repository/pypi-group/simple/'
priority = "supplemental"


[[tool.poetry.source]]
name = 'pytorch-cu118'
url = 'https://download.pytorch.org/whl/cu118'
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.6"

[tool.poetry.group.test.dependencies]
pytest = "7.3.1"
pytest-asyncio = "0.21.0"
matplotlib = "3.7.1"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend" # wrapper around poetry.core.masonry.api

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"